@@ -1,6 +1,9 @@
+# frozen_string_literal: true
+
 source "https://rubygems.org"
 
-ruby "~> 2.4"
+ruby "~> 2.5"
 
-gem "pry", "~> 0.12.2"
-gem "assert", "~> 2.18.1"
+gem "pry"
+gem "assert",           "~> 2.19.3"
+gem "much-style-guide", "~> 0.6.0"
