@@ -1,5 +1,6 @@
-require "assert"
+# frozen_string_literal: true
 
+require "assert"
 require "cli"
 
 class CLIRB
@@ -8,7 +9,7 @@ class CLIRB
     setup do
       @cli = CLIRB.new
     end
-    subject { @cli }
+    subject{ @cli }
 
     should have_readers :args, :opts, :data
     should have_imeths  :option, :parse!, :to_s
@@ -23,28 +24,28 @@ class CLIRB
     end
 
     should "raise `VersionExit` parsing `--version`" do
-      assert_raises(CLIRB::VersionExit) { cli_parse subject, "--version" }
+      assert_raises(CLIRB::VersionExit){ cli_parse subject, "--version" }
     end
 
     should "raise `HelpExit` when parsing `--help`" do
-      assert_raises(CLIRB::HelpExit) { cli_parse subject, "--help" }
+      assert_raises(CLIRB::HelpExit){ cli_parse subject, "--help" }
     end
 
     should "parse the args, opts, and full data" do
-      cli = CLIRB.new{ option "verbose", "verbosity"}
+      cli = CLIRB.new{ option "verbose", "verbosity" }
       cli_parse(cli, "an", "arg", "-v")
 
       assert_equal ["an", "arg"], cli.args
       assert_kind_of Hash, cli.opts
       assert_equal 1, cli.opts.size
-      assert_equal cli.args+[cli.opts], cli.data
+      assert_equal cli.args + [cli.opts], cli.data
     end
   end
 
   class SwitchTests < UnitTests
     desc "when parsing a switch opt"
     setup do
-      @cli = CLIRB.new{ option "verbose", "verbosity"}
+      @cli = CLIRB.new{ option "verbose", "verbosity" }
     end
 
     should "default to nil" do
@@ -71,7 +72,7 @@ class CLIRB
   class SingleValueTests < UnitTests
     desc "when parsing a single value opt"
     setup do
-      @cli = CLIRB.new{ option "skill", "skillz", :value => "" }
+      @cli = CLIRB.new{ option "skill", "skillz", value: "" }
     end
 
     should "set the default" do
@@ -90,7 +91,7 @@ class CLIRB
     end
 
     should "type-cast the value" do
-      cli = CLIRB.new{ option "skill", "skillz", :value => 1 }
+      cli = CLIRB.new{ option "skill", "skillz", value: 1 }
       cli.parse! ["-s", "12"]
       assert_equal 12, cli.opts["skill"]
     end
@@ -99,7 +100,7 @@ class CLIRB
   class ListValueTests < UnitTests
     desc "when parsing a list value opt"
     setup do
-      @cli = CLIRB.new{ option :skill, "skillz", :value => [] }
+      @cli = CLIRB.new{ option :skill, "skillz", value: [] }
     end
 
     should "set the list values by parsing the value as comma-separated" do
